requirements.txt
langchain
langchain-community
langchain-openai
langchain-chroma
chromadb
openai
python-dotenv
